% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-adaptive_spline.R
\name{adaptive_spline_stan_fit_impl}
\alias{adaptive_spline_stan_fit_impl}
\title{Low-Level ARIMA function for translating modeltime to forecast}
\usage{
adaptive_spline_stan_fit_impl(
  x,
  y,
  degree = 1,
  maxInt = 3,
  maxInt.cat = 3,
  npart = NULL,
  ...
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{degree}{degree of splines}

\item{maxInt}{integer for maximum degree of interaction in spline basis functions}

\item{maxInt.cat}{(categorical input only) integer for maximum degree of interaction of categorical inputs}

\item{npart}{minimum number of non-zero points in a basis function}

\item{...}{Extra arguments}
}
\value{
A modeltime model
}
\description{
Low-Level ARIMA function for translating modeltime to forecast
}
