% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-bayesian_structural_reg.R
\name{bayesian_structural_stan_fit_impl}
\alias{bayesian_structural_stan_fit_impl}
\title{Low-Level ARIMA function for translating modeltime to forecast}
\usage{
bayesian_structural_stan_fit_impl(formula, data, family = "gaussian", ...)
}
\arguments{
\item{formula}{A dataframe of xreg (exogenous regressors)}

\item{data}{A numeric vector of values to fit}

\item{family}{The model family for the observation equation.
Non-Gaussian model families use data augmentation to recover a conditionally Gaussian model.}

\item{...}{Additional arguments passed to \code{forecast::Arima}}
}
\value{
A modeltime model
}
\description{
Low-Level ARIMA function for translating modeltime to forecast
}
