% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-exp_smoothing_params.R
\name{exponential_smoothing_params}
\alias{exponential_smoothing_params}
\alias{seasonality_type}
\alias{method}
\alias{error_method}
\title{Tuning Parameters for Exponential Smoothing Models}
\usage{
seasonality_type()

method()

error_method()
}
\value{
A parameter

A parameter

A parameter
}
\description{
Tuning Parameters for Exponential Smoothing Models
}
\details{
The main parameters for Exponential Smoothing models are:
\itemize{
\item \code{garch_order}: Integer with the garch order.
\item \code{arch_order}: Integer with the arch_order.
\item \code{mgarch_order}: Integer with the mgarch order.
\item \code{garch_t_student}: A boolean value to specify for a generalized t-student garch model.
\item \code{asymmetry}: a string value for the asymmetric function for an asymmetric GARCH process. By default the
value "none" for standard GARCH process. If "logit" a logistic function is used for asymmetry, and if
"exp" an exponential function is used.
\item \code{non_seasonal_ar}: The order of the non-seasonal auto-regressive (AR) terms.
\item \code{non_seasonal_ma}: The order of the non-seasonal moving average (MA) terms.
\item \code{markov_chains}: The number of markov chains.
\item \code{adapt_delta}: The thin of the jumps in a HMC method
\item \code{tree_depth}: Maximum depth of the trees
}
}
\examples{
non_seasonal_ar()

non_seasonal_differences()

non_seasonal_ma()


}
