% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-garch_multivariate_reg.R
\name{gogarch_rmgarch_multi_fit_impl}
\alias{gogarch_rmgarch_multi_fit_impl}
\title{Low-Level GARCH function for translating modeltime to forecast}
\usage{
gogarch_rmgarch_multi_fit_impl(
  formula,
  data,
  spec_type = NULL,
  period = "auto",
  ...
)
}
\arguments{
\item{formula}{A dataframe of xreg (exogenous regressors)}

\item{data}{A numeric vector of values to fit}

\item{spec_type}{NA}

\item{period}{auto}

\item{...}{Additional arguments passed to \code{forecast::Arima}}
}
\value{
A fitted model
}
\description{
Low-Level GARCH function for translating modeltime to forecast
}
