% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-garch_multivariate_reg.R
\name{gogarch_rmgarch_multi_predict_impl}
\alias{gogarch_rmgarch_multi_predict_impl}
\title{Bridge prediction function for GARCH models}
\usage{
gogarch_rmgarch_multi_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{stats::predict()}}
}
\value{
A nested tibble
}
\description{
Bridge prediction function for GARCH models
}
