% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-naive_params.R
\name{naive_params}
\alias{naive_params}
\alias{seasonal_random_walk}
\title{Tuning Parameters for Random Walk Models}
\usage{
seasonal_random_walk()
}
\value{
A parameter
}
\description{
Tuning Parameters for Random Walk Models
}
\details{
The main parameters for Random Walk Models are:
\itemize{
\item \code{seasonal_random_walk}: A boolean value for select a seasonal random walk instead.
\item \code{markov_chains}: The number of markov chains.
\item \code{adapt_delta}: The thin of the jumps in a HMC method
\item \code{tree_depth}: Maximum depth of the trees
}
}
