% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_modeltime_bridge.R
\name{new_modelgarch_bridge}
\alias{new_modelgarch_bridge}
\title{Constructor for creating garchmodels models}
\usage{
new_modelgarch_bridge(class, models, data, extras = NULL, desc = NULL)
}
\arguments{
\item{class}{A class name that is used for creating custom printing messages}

\item{models}{A list containing one or more models}

\item{data}{A data frame (or tibble) containing 4 columns:
(date column with name that matches input data), .actual, .fitted, and .residuals.}

\item{extras}{An optional list that is typically used for transferring preprocessing recipes
to the predict method.}

\item{desc}{An optional model description to appear when printing your modeltime objects}
}
\value{
A list with the constructor
}
\description{
These functions are used to construct new \code{garchmodels} bridge functions that
connect the \code{tidymodels} infrastructure to time-series models containing date or date-time features.
}
