% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rIBM}
\alias{rIBM}
\title{The IBM's Daily Returns}
\format{
A \code{tibble} with 3523 rows and 2 variables:
\itemize{
\item \code{date} Date. Timestamp information. Daily format.
\item \code{daily_returns} Numeric. Value at the corresponding timestamp.
}
}
\usage{
rIBM
}
\description{
The IBM's Daily Returns
}
\examples{
rIBM


}
\keyword{datasets}
