% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rX_longer}
\alias{rX_longer}
\title{IBM, Google and BP's daily returns in long format}
\format{
A \code{tibble} with 8550 rows and 3 variables:
\itemize{
\item \code{id} Factor. Unique series identifier
\item \code{date} Date. Timestamp information. Daily format.
\item \code{value} Numeric. Value at the corresponding timestamp.
}
}
\usage{
rX_longer
}
\description{
IBM, Google and BP's daily returns in long format
}
\details{
\if{html}{\out{<div class="r">}}\preformatted{library(timetk)
m750_splits <- time_series_split(m750, assess = "2 years", cumulative = TRUE)
}\if{html}{\out{</div>}}
}
\examples{

rX_longer

}
\keyword{datasets}
