% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip_sarima_reg.R
\name{Sarima_stan_fit_impl}
\alias{Sarima_stan_fit_impl}
\title{Low-Level ARIMA function for translating modeltime to forecast}
\usage{
Sarima_stan_fit_impl(
  x,
  y,
  period = "auto",
  p = 0,
  d = 0,
  q = 0,
  P = 0,
  D = 0,
  Q = 0,
  chains = 4,
  iter = 2000,
  warmup = iter/2,
  adapt.delta = 0.9,
  tree.depth = 10,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{period}{A seasonal frequency. Uses "auto" by default. A character phrase
of "auto" or time-based phrase of "2 weeks" can be used if a date or date-time variable is provided.}

\item{p}{The order of the non-seasonal auto-regressive (AR) terms. Often denoted "p" in pdq-notation.}

\item{d}{The order of integration for non-seasonal differencing. Often denoted "d" in pdq-notation.}

\item{q}{The order of the non-seasonal moving average (MA) terms. Often denoted "q" in pdq-notation.}

\item{P}{The order of the seasonal auto-regressive (SAR) terms. Often denoted "P" in PDQ-notation.}

\item{D}{The order of integration for seasonal differencing. Often denoted "D" in PDQ-notation.}

\item{Q}{The order of the seasonal moving average (SMA) terms. Often denoted "Q" in PDQ-notation.}

\item{chains}{An integer of the number of Markov Chains chains to be run, by default 4 chains are run.}

\item{iter}{An integer of total iterations per chain including the warm-up, by default the number of iterations are 2000.}

\item{warmup}{A positive integer specifying number of warm-up (aka burn-in) iterations. This also specifies the number of iterations used for step-size adaptation, so warm-up samples should not be used for inference. The number of warmup should not be larger than iter and the default is iter/2.}

\item{adapt.delta}{An optional real value between 0 and 1, the thin of the jumps in a HMC method. By default is 0.9}

\item{tree.depth}{An integer of the maximum depth of the trees evaluated during each iteration. By default is 10.}

\item{seed}{An integer with the seed for using when predicting with the model.}

\item{...}{Additional arguments passed to \code{forecast::Arima}}
}
\value{
A modeltime model
}
\description{
Low-Level ARIMA function for translating modeltime to forecast
}
