% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-ssm_params.R
\name{ssm_params}
\alias{ssm_params}
\alias{trend_model}
\alias{damped_model}
\alias{seasonal_model}
\title{Tuning Parameters for Additive Linear State Space Regression Models}
\usage{
trend_model()

damped_model()

seasonal_model()
}
\value{
A parameter

A parameter

A parameter
}
\description{
Tuning Parameters for Additive Linear State Space Regression Models
}
\details{
The main parameters for Additive Linear State Space Regression Models are:
\itemize{
\item \code{trend_model}: A boolean value to specify a trend local level model.
\item \code{damped_model}: A boolean value to specify a damped trend local level model.
\item \code{seasonal_model}: A boolean value to specify a seasonal trend local level model.
\item \code{markov_chains}: The number of markov chains.
\item \code{adapt_delta}: The thin of the jumps in a HMC method
\item \code{tree_depth}: Maximum depth of the trees
}
}
\examples{

damped_model()

seasonal_model()


}
